+++ POSSystem/ViewModels/EmployeeViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using POSSystem.Models;
-using POSSystem.Repository;
+using POSSystem.Repositories;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
