+++ POSSystem/ViewModels/RegisterViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using POSSystem.Models;
-using POSSystem.Repository;
+using POSSystem.Repositories;
 using System;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
